requirements-dev.txt
-r requirements.txt
pytest-django>=4.7
pytest-cov>=4.1
coverage>=7.0
black>=24.0
flake8>=7.0
pip-audit>=2.7
bandit[toml]>=1.7
